Move ConnectionListModel into its own source file
