+++ core/CMakeLists.txt
@@ -91,6 +91,7 @@ set(SUBSURFACE_CORE_LIB_SRCS
 	gpslocation.cpp
 	cloudstorage.cpp
 	downloadfromdcthread.cpp
+	connectionlistmodel.cpp
 
 	#Subsurface Qt have the Subsurface structs QObjectified for easy access via QML.
 	subsurface-qt/DiveObjectHelper.cpp
