+++ core/btdiscovery.h
@@ -21,22 +21,6 @@
 void saveBtDeviceInfo(const char* devaddr, QBluetoothDeviceInfo deviceInfo);
 QBluetoothDeviceInfo getBtDeviceInfo(const char* devaddr);
 
-class ConnectionListModel : public QAbstractListModel {
-	Q_OBJECT
-public:
-	enum CLMRole {
-		AddressRole = Qt::UserRole + 1
-	};
-	ConnectionListModel(QObject *parent = 0);
-	QHash<int, QByteArray> roleNames() const;
-	QVariant data(const QModelIndex &index, int role = AddressRole) const;
-	QString address(int idx) const;
-	int rowCount(const QModelIndex &parent = QModelIndex()) const;
-	void addAddress(const QString address);
-private:
-	QStringList m_addresses;
-};
-
 class BTDiscovery : public QObject {
 	Q_OBJECT
